Fixed bugs, found bugs, made bugs. See comments on this commit.
Query sometimes finds nodes that are referencing themselves. Not sure if it is in the file IO or in the tree. The tree isn't storing all of the nodes it needs to at the moment. It loses some. First thing when I wake up tomorrow I will make a test to see if a node references its self in the btree construction, but I haven't seen any infitite loops on that part, so probably not. Going to get some sleep, lots more to do tomorrow.